feat: move ebay inventory fields to dedicated tab
