+++ src/core/integrations/integrations/integrations-show/containers/inventory-fields/InventoryFields.vue
@@ -1,0 +1,27 @@
+<script setup lang="ts">
+import { computed } from 'vue';
+import { IntegrationTypes } from '../../../integrations';
+import { EbayInventoryProperties } from '../properties/containers/ebay-properties';
+
+const props = defineProps<{ id: string; salesChannelId: string; type: string }>();
+const emit = defineEmits(['pull-data']);
+
+const currentComponent = computed(() => {
+  switch (props.type) {
+    case IntegrationTypes.Ebay:
+      return EbayInventoryProperties;
+    default:
+      return null;
+  }
+});
+</script>
+
+<template>
+  <component
+    v-if="currentComponent"
+    :is="currentComponent"
+    :id="id"
+    :sales-channel-id="salesChannelId"
+    @pull-data="emit('pull-data')"
+  />
+</template>
