+++ src/core/integrations/integrations/integrations-show/containers/inventory-fields/index.ts
@@ -1,0 +1,1 @@
+export { default as InventoryFields } from './InventoryFields.vue';
