+++ src/core/integrations/integrations/integrations-show/containers/properties/containers/ebay-properties/index.ts
@@ -1,1 +1,2 @@
 export { default as EbayProperties } from './EbayProperties.vue';
+export { default as EbayInventoryProperties } from './EbayInventoryProperties.vue';
